#1218 split up tables into files and buttons
